# docs/openapi.yaml
openapi: 3.0.1
info:
  description: "REST API for managing gym class sessions, memberships, bookings and\
    \ waitlists."
  title: Fitness Class Booking System
  version: "1.0"
servers:
- description: Default server
  url: /
...
paths:
  /api/v1/bookings:
    post:
      description: "Creates a new booking for a member on a session. Returns 201 on\
        \ success, 200 on idempotent replay. Rules: suspension checked first, credits\
        \ checked before capacity. Supports optional Idempotency-Key header for safe\
        \ retries."
      operationId: createBooking
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: Idempotent replay of previous successful booking (no Location
            header)
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: Booking successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Malformed JSON request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Member suspended or membership inactive
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Member or session not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: "Conflict: session full/cancelled/started, duplicate booking,\
            \ no credits, overlapping booking, idempotency key mismatch"
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Validation error or unknown field
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Create a booking
      tags:
      - Bookings
  /api/v1/bookings/{id}:
    delete:
      description: Cancels a booked spot. A successful cancellation does not imply
        available capacity because a waitlisted member may be promoted immediately.
        Credits are refunded when cancellation is at least four hours before the session.
      operationId: cancelBooking
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
          description: Booking cancelled
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
          description: BOOKING_NOT_FOUND
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
          description: BOOKING_NOT_CANCELLABLE or SESSION_ALREADY_STARTED
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CancelBookingResponse"
          description: Invalid booking ID or reason exceeds 255 characters
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Cancel a booking
      tags:
      - Bookings
    get:
      description: Returns the full booking representation with session detail.
      operationId: getBooking
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: Booking found
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: BOOKING_NOT_FOUND
        "422":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
          description: Malformed UUID
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal server error
      summary: Get a booking
      tags:
      - Bookings
...
components:
  schemas:
...
    BookingResponse:
      type: object
      properties:
        bookedAt:
          type: string
        cancellationType:
          type: string
        checkedInAt:
          type: string
          format: date-time
        checkedInBy:
          type: string
        createdAt:
          type: string
        creditDeducted:
          type: boolean
        creditRefunded:
          type: boolean
        id:
          type: string
        memberId:
          type: string
        membershipId:
          type: string
        session:
          $ref: "#/components/schemas/SessionDetail"
        sessionId:
          type: string
        source:
          type: string
        status:
          type: string
        updatedAt:
          type: string
    CancelBookingRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 255
          minLength: 0
    CancelBookingResponse:
      type: object
      properties:
        bookingId:
          type: string
          format: uuid
        cancellationType:
          type: string
        cancelledAt:
          type: string
          format: date-time
        creditRefunded:
          type: boolean
        promotedWaitlistEntryId:
          type: string
          format: uuid
        status:
          type: string
...
    CreateBookingRequest:
      type: object
      properties:
        memberId:
          type: string
        sessionId:
          type: string
      required:
      - memberId
      - sessionId
...
    ErrorEnvelope:
      type: object
      properties:
        code:
          type: string
          enum:
          - MALFORMED_REQUEST
          - NOT_PERMITTED
          - MEMBER_SUSPENDED
          - MEMBER_INACTIVE
          - NOT_FOUND
          - MEMBER_NOT_FOUND
          - MEMBERSHIP_PLAN_NOT_FOUND
          - MEMBERSHIP_NOT_FOUND
          - CLASS_TYPE_NOT_FOUND
          - INSTRUCTOR_NOT_FOUND
          - ROOM_NOT_FOUND
          - SESSION_NOT_FOUND
          - BOOKING_NOT_FOUND
          - WAITLIST_ENTRY_NOT_FOUND
          - CONFLICT
          - CONCURRENT_MODIFICATION
          - DUPLICATE_BOOKING
          - SESSION_FULL
          - SESSION_CANCELLED
          - SESSION_ALREADY_CANCELLED
          - SESSION_NOT_BOOKABLE
          - BOOKING_ALREADY_CANCELLED
          - BOOKING_NOT_CANCELLABLE
          - BOOKING_ALREADY_CHECKED_IN
          - CHECK_IN_WINDOW_NOT_OPEN
          - WAITLIST_ALREADY_JOINED
          - WAITLIST_ALREADY_PROCESSED
          - MEMBERSHIP_ALREADY_ACTIVE
          - CREDITS_INSUFFICIENT
          - MEMBERSHIP_NO_CREDITS
          - OVERLAPPING_BOOKING
          - BOOKING_OVERLAPS_EXISTING
          - LATE_CANCEL_NO_REFUND
          - INSTRUCTOR_SCHEDULE_CONFLICT
          - ROOM_SCHEDULE_CONFLICT
          - CLASS_TYPE_NAME_ALREADY_EXISTS
          - CLASS_TYPE_INACTIVE
          - INSTRUCTOR_EMAIL_ALREADY_EXISTS
          - INSTRUCTOR_INACTIVE
          - INSTRUCTOR_HAS_FUTURE_SESSIONS
          - ROOM_NAME_ALREADY_EXISTS
          - ROOM_INACTIVE
          - ROOM_HAS_FUTURE_SESSIONS
          - SESSION_INSTRUCTOR_CONFLICT
          - SESSION_ROOM_CONFLICT
          - SESSION_CAPACITY_EXCEEDS_ROOM
          - IDEMPOTENCY_KEY_CONFLICT
          - JOB_ALREADY_RUNNING
          - MEMBER_EMAIL_ALREADY_EXISTS
          - MEMBER_ALREADY_SUSPENDED
          - MEMBER_NOT_SUSPENDED
          - PLAN_NAME_ALREADY_EXISTS
          - PLAN_ALREADY_INACTIVE
          - PLAN_INACTIVE
          - MEMBERSHIP_ALREADY_QUEUED
          - MEMBERSHIP_START_BEFORE_CURRENT_EXPIRY
          - MEMBERSHIP_NOT_CANCELLABLE
          - SESSION_NOT_EDITABLE
          - SESSION_ALREADY_STARTED
          - SESSION_CAPACITY_BELOW_BOOKED
          - VALIDATION_FAILED
          - OUT_OF_RANGE
          - INVALID_DATE_RANGE
          - INVALID_PAGINATION
          - INVALID_SORT_FIELD
          - INVALID_FORMAT
          - INVALID_ENUM
          - UNKNOWN_FIELD
          - FUTURE_REQUIRED
          - INVALID_RANGE
          - RECURRENCE_EMPTY
          - INTERNAL_ERROR
        detail:
          type: string
        errors:
          type: array
          description: Per-field validation errors. Present ONLY on 422 responses.
          items:
            $ref: "#/components/schemas/FieldError"
        instance:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        title:
          type: string
        traceId:
          type: string
        type:
          type: string
    FieldError:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
        rejectedValue:
          type: object
...
    SessionDetail:
      type: object
      properties:
        classTypeName:
          type: string
        endsAt:
          type: string
          format: date-time
        id:
          type: string
        instructorName:
          type: string
        roomName:
          type: string
        startsAt:
          type: string
          format: date-time
...